+++ rust/lance-table/src/transaction/index_maintenance.rs
@@ -18,10 +18,9 @@ use lance_core::datatypes::Schema;
 use lance_core::{Error, Result};
 use roaring::RoaringBitmap;
 use std::collections::{HashMap, HashSet};
-use uuid::Uuid;
 
 impl Transaction {
-    pub(crate) fn register_pure_rewrite_rows_update_frags_in_indices(
+    pub(super) fn register_pure_rewrite_rows_update_frags_in_indices(
         indices: &mut [IndexMetadata],
         pure_update_frag_ids: &[u64],
         original_fragment_ids: &[u64],
@@ -84,7 +83,7 @@ impl Transaction {
 
     /// If an operation modifies one or more fields in a fragment then we need to remove
     /// that fragment from any indices that cover one of the modified fields.
-    pub(crate) fn prune_updated_fields_from_indices(
+    pub(super) fn prune_updated_fields_from_indices(
         indices: &mut [IndexMetadata],
         updated_fragments: &[Fragment],
         fields_modified: &[u32],
@@ -130,7 +129,7 @@ impl Transaction {
     /// only drops indices for *removed* fields, so without this the index keeps
     /// covering the rewritten fragments with stale entries. Remove each such fragment
     /// from any index covering a field whose backing data file changed.
-    pub(crate) fn prune_merge_rewritten_fields_from_indices(
+    pub(super) fn prune_merge_rewritten_fields_from_indices(
         indices: &mut [IndexMetadata],
         prev_fragments: &[Fragment],
         new_fragments: &[Fragment],
@@ -171,7 +170,7 @@ impl Transaction {
     /// fragment no longer signals that staleness to the query path. Drop each
     /// rewritten (new) fragment from the coverage of any index covering a field
     /// such an overlay supplied, so those rows fall back to a flat scan.
-    pub(crate) fn prune_overlay_stale_fields_from_indices(
+    pub(super) fn prune_overlay_stale_fields_from_indices(
         indices: &mut [IndexMetadata],
         groups: &[RewriteGroup],
     ) {
@@ -222,7 +221,7 @@ impl Transaction {
         }
     }
 
-    pub(crate) fn retain_relevant_indices(
+    pub(super) fn retain_relevant_indices(
         indices: &mut Vec<IndexMetadata>,
         schema: &Schema,
         fragments: &[Fragment],
@@ -321,7 +320,7 @@ impl Transaction {
         });
     }
 
-    pub(crate) fn recalculate_fragment_bitmap(
+    pub(super) fn recalculate_fragment_bitmap(
         old: &RoaringBitmap,
         groups: &[RewriteGroup],
     ) -> Result<RoaringBitmap> {
@@ -355,7 +354,7 @@ impl Transaction {
         Ok(new_bitmap)
     }
 
-    pub(crate) fn handle_rewrite_indices(
+    pub(super) fn handle_rewrite_indices(
         indices: &mut [IndexMetadata],
         rewritten_indices: &[RewrittenIndex],
         groups: &[RewriteGroup],
@@ -396,7 +395,7 @@ impl Transaction {
         Ok(())
     }
 
-    pub(crate) fn handle_rewrite_fragments(
+    pub(super) fn handle_rewrite_fragments(
         final_fragments: &mut Vec<Fragment>,
         groups: &[RewriteGroup],
         fragment_id: &mut u64,
@@ -461,6 +460,7 @@ impl Transaction {
 mod tests {
     use super::*;
     use crate::transaction::test_support::overlay_with_field;
+    use uuid::Uuid;
 
     #[test]
     fn test_rewrite_fragments() {
